+++ ChillaxScraps/Plugin.cs
@@ -32,6 +32,7 @@ namespace ChillaxScraps
                 case 3: script = item.spawnPrefab.AddComponent<Food>(); break;
                 case 4: script = item.spawnPrefab.AddComponent<UnoReverse>(); break;
                 case 5: script = item.spawnPrefab.AddComponent<EmergencyMeeting>(); break;
+                case 6: script = item.spawnPrefab.AddComponent<SuperSneakers>(); break;
                 default: return;
             }
             script.grabbable = true;
@@ -60,7 +61,9 @@ namespace ChillaxScraps
                 bundle.LoadAsset<AudioClip>(directory + "_audio/Boink.wav"),
                 bundle.LoadAsset<AudioClip>(directory + "_audio/EatSFX.wav"),
                 bundle.LoadAsset<AudioClip>(directory + "_audio/uno-reverse-biaatch.wav"),
-                bundle.LoadAsset<AudioClip>(directory + "_audio/emergencymeeting.wav")
+                bundle.LoadAsset<AudioClip>(directory + "_audio/emergencymeeting.wav"),
+                bundle.LoadAsset<AudioClip>(directory + "_audio/sneakers-activate.wav"),
+                bundle.LoadAsset<AudioClip>(directory + "_audio/sneakers-deactivate.wav")
             };
 
             var scraps = new List<Scrap> {
@@ -71,7 +74,8 @@ namespace ChillaxScraps
                 new Scrap("Moai/MoaiItem.asset", 9),
                 new Scrap("UnoReverseCard/UnoReverseCardItem.asset", 8, 4),
                 new Scrap("FroggyChair/FroggyChairItem.asset", 10),
-                new Scrap("EmergencyMeeting/EmergencyMeetingItem.asset", 6, 5)
+                new Scrap("EmergencyMeeting/EmergencyMeetingItem.asset", 6, 5),
+                new Scrap("SuperSneakers/SuperSneakersItem.asset", 10, 6)
             };
 
             int i = 0; config = new Config(base.Config, scraps);